main.py
"""
主程序入口 - 整合所有模块的示例
"""
import yaml
import logging
from pathlib import Path
import torch
import torch.nn as nn

from src.data_processing.processor import DataProcessor
from src.scraper.web_scraper import WebScraper
from src.model.trainer import ModelTrainer, SimpleCNN
from src.optimization.optimizer import ModelOptimizer
from src.multimodal.recognizer import MultimodalRecognizer
from src.deployment.server import ModelDeploymentManager

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


def load_config(config_path: str = "config.yaml"):
    """加载配置文件"""
    with open(config_path, 'r', encoding='utf-8') as f:
        config = yaml.safe_load(f)
    return config


def main():
    """主函数"""
    # 加载配置
    config = load_config()
    logger.info("配置加载完成")
    
    # 1. 数据处理示例
    logger.info("=" * 50)
    logger.info("1. 数据处理模块示例")
    logger.info("=" * 50)
    
    processor = DataProcessor(data_dir=config['data']['data_dir'])
    # 数据处理示例代码...
    
    # 2. Web爬虫示例
    logger.info("=" * 50)
    logger.info("2. Web爬虫模块示例")
    logger.info("=" * 50)
    
    scraper = WebScraper(
        delay=config['scraper']['delay'],
        timeout=config['scraper']['timeout']
    )
    # 爬虫示例代码...
    
    # 3. 模型训练示例
    logger.info("=" * 50)
    logger.info("3. 模型训练模块示例")
    logger.info("=" * 50)
    
    # 创建示例模型
    model = SimpleCNN(vocab_size=10000, embed_dim=128, num_classes=2)
    trainer = ModelTrainer(model, device=config['training']['device'])
    # 训练示例代码...
    
    # 4. 模型优化示例
    logger.info("=" * 50)
    logger.info("4. 模型优化模块示例")
    logger.info("=" * 50)
    
    optimizer = ModelOptimizer(model)
    
    # 优化模型
    optimized_model = optimizer.optimize(
        quantization=config['optimization']['quantization']['enabled'],
        pruning=config['optimization']['pruning']['enabled'],
        pruning_ratio=config['optimization']['pruning']['ratio'],
        quantization_bits=config['optimization']['quantization']['bits']
    )
    
    # 获取模型大小
    model_size = optimizer.get_model_size(optimized_model)
    logger.info(f"模型大小: {model_size}")
    
    # 导出ONNX
    dummy_input = torch.randn(1, 128)
    onnx_export_path = "models/optimized_model.onnx"
    optimizer.export_to_onnx(
        optimized_model,
        dummy_input,
        onnx_export_path
    )

    # 导出的ONNX模型可以用于多种生产部署：
    # 1. 使用 ONNX Runtime 进行高效推理:
    #    import onnxruntime as ort
    #    ort_session = ort.InferenceSession(onnx_export_path)
    #    input_data = np.random.randn(1, 128).astype(np.float32)
    #    outputs = ort_session.run(None, {'input': input_data})
    #
    # 2. 集成到 FastAPI/Flask 服务，作为线上推理后端
    #    只需在API接口中读取ONNX模型并用onnxruntime推理即可
    #
    # 3. 部署到云原生平台：
    #    - TensorRT（NVIDIA GPU推理优化）
    #    - OpenVINO（Intel平台推理）
    #    - Triton Inference Server
    #
    # 4. 跨语言调用（如C++/Java）：
    #    ONNX模型支持多种语言的推理SDK，方便嵌入各类业务系统

    logger.info(f"ONNX模型已保存，支持ONNX Runtime等多后端部署: {onnx_export_path}")
    
    # 5. 多模态识别示例
    logger.info("=" * 50)
    logger.info("5. 多模态识别模块示例")
    logger.info("=" * 50)
    
    multimodal_recognizer = MultimodalRecognizer()
    # 多模态识别示例代码...
    
    # 6. 模型部署示例
    logger.info("=" * 50)
    logger.info("6. 模型部署模块示例")
    logger.info("=" * 50)
    
    deployment_manager = ModelDeploymentManager(model, "example_model")
    
    # 导出模型
    deployment_manager.export_model("models/exported_model.pt", format="torchscript")
    
    logger.info("=" * 50)
    logger.info("所有模块示例完成！")
    logger.info("=" * 50)


if __name__ == "__main__":
    main()

