requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web爬虫
requests>=2.31.0
beautifulsoup4>=4.12.0
scrapy>=2.11.0
selenium>=4.15.0

# 模型优化
onnx>=1.15.0
onnxruntime>=1.16.0
onnxsim>=0.4.0

# 模型量化
pytorch-quantization>=2.1.0

# 多模态处理
opencv-python>=4.8.0
Pillow>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0

# 模型部署
flask>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.7.0

# 可视化
matplotlib>=3.8.0
seaborn>=0.13.0

# 工具库
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 大模型相关
transformers>=4.35.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

